2003-06-21  Alexander V. Lukyanov <lav@yars.free.net>
	* FileCopy.cc: optimize FileCopyPeerFDStream speed by writing out
	  data by portions of >=PUT_LL_MIN or by timeout 0.1 second.
	* buffer.cc, buffer.h: add put_ll_timer in IOBufferFDStream too.
	* FileSet.cc: remove unneeded type casting.
	* GetJob.cc: create FileCopyPeerFDStream for file: url scheme.
	  This fixes closing and re-opening output file when transfer
	  breaks and restarts. Besides, it is slightly faster.
	* Http.cc: turn off sending plain password in URI along with
	  Authorization in hftp.
	* Timer.h: add Stop method.
	* ftpclass.cc, ftpclass.h, resource.cc: new settings ftp:use-size
	  and ftp:use-mdtm.

+++ src/FileSet.cc
@@ -347,7 +347,7 @@ bool  FileInfo::SameAs(const FileInfo *fi,int ignore)
       if(p<fi->date_prec)
 	 p=fi->date_prec;
       if(!(ignore&IGNORE_DATE_IF_OLDER && date<fi->date)
-      && labs((long)date-(long)(fi->date))>p)
+      && labs(date-fi->date)>p)
 	 return false;
    }
